+++ hugegraph-client-go/api/v1/version_test.go
@@ -15,24 +15,26 @@
  * under the License.
  */
 
-
-package schema_test
+package v1_test
 
 import (
-    "fmt"
-    "github.com/apache/incubator-hugegraph-toolchain/hugegraph-client-go"
-    "log"
     "testing"
+
+    hugegraph "github.com/apache/incubator-hugegraph-toolchain/hugegraph-client-go"
 )
 
-func TestSchemaRequest_Do(t *testing.T) {
+func TestVersionRequest_Do(t *testing.T) {
+
     client, err := hugegraph.NewDefaultCommonClient()
     if err != nil {
-        log.Println(err)
+        t.Errorf("NewDefaultCommonClient() error = %v", err)
     }
-    schema, err := client.Schema()
+
+    resp, err := client.Version()
     if err != nil {
-        log.Fatalln(err)
+        t.Errorf(" client.Version() error = %v", err)
     }
-    fmt.Println(schema.Data)
+    if resp.StatusCode != 200 {
+        t.Errorf(" client.Version() code error = %v", resp.StatusCode)
+    }
 }
